# atmSystem/OpenAPI Yaml Files/Service api-docs.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Atm Project API",
    "description": "Atm Project API Service",
    "version": "0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticatedUserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "operationId": "loginUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticatedUserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/users/withdraw": {
      "patch": {
        "tags": [
          "user-controller"
        ],
        "operationId": "createUserWithdraw",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BasicTransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/users/transfer": {
      "patch": {
        "tags": [
          "user-controller"
        ],
        "operationId": "createUserTransfer",
        "parameters": [
          {
            "name": "idAccount1",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idAccount2",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/users/deposit": {
      "patch": {
        "tags": [
          "user-controller"
        ],
        "operationId": "createUserDeposit",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BasicTransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getUserDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          }
        }
      }
    },
    "/users/statement": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getBankStatementFromAPeriod",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bank/users": {
      "get": {
        "tags": [
          "bank-controller"
        ],
        "operationId": "getUsersList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankUserDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bank/user-most-transactions": {
      "get": {
        "tags": [
          "bank-controller"
        ],
        "operationId": "getBankUserWithMostTransactions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BankUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/bank/transactions": {
      "get": {
        "tags": [
          "bank-controller"
        ],
        "operationId": "getTransactionList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bank/transactions-between-dates": {
      "get": {
        "tags": [
          "bank-controller"
        ],
        "operationId": "getTransactionsBetweenDates",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bank/highest-balance": {
      "get": {
        "tags": [
          "bank-controller"
        ],
        "operationId": "getAccountWithHighestBalance",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          }
        }
      }
    },
    "/bank/date-most-transactions": {
      "get": {
        "tags": [
          "bank-controller"
        ],
        "operationId": "getDateUserWithMostTransactions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      }
    },
    "/bank/balance": {
      "get": {
        "tags": [
          "bank-controller"
        ],
        "operationId": "getBankBalance",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
    "/bank/accounts": {
      "get": {
        "tags": [
          "bank-controller"
        ],
        "operationId": "getUsersAccountList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/close-account/{id}": {
      "delete": {
        "tags": [
          "user-controller"
        ],
        "operationId": "closeBankAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthenticatedUserDto": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "BasicTransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "number"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "DEPOSIT",
              "WITHDRAW",
              "TRANSFER"
            ]
          }
        }
      },
      "TransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "number"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "DEPOSIT",
              "WITHDRAW",
              "TRANSFER"
            ]
          },
          "fromIdAccount": {
            "type": "integer",
            "format": "int64"
          },
          "toIdAccount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "balance": {
            "type": "number"
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BasicTransactionDto"
            }
          }
        }
      },
      "BankUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "bankUser": {
            "$ref": "#/components/schemas/BankUser"
          },
          "balance": {
            "type": "number"
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        }
      },
      "BankUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "number"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "DEPOSIT",
              "WITHDRAW",
              "TRANSFER"
            ]
          },
          "fromIdAccount": {
            "type": "integer",
            "format": "int64"
          },
          "toIdAccount": {
            "type": "integer",
            "format": "int64"
          },
          "account": {
            "$ref": "#/components/schemas/Account"
          }
        }
      }
    }
  }
}