# atmSystem/src/main/java/com/atm/dto/AccountDto.java
package com.atm.dto;

import lombok.Getter;
import lombok.Setter;

import java.math.BigDecimal;
import java.util.List;

@Getter
@Setter
public class AccountDto {
    private Long id;
    private BigDecimal balance;
    private List<BasicTransactionDto> transactions;

}