# atmSystem/src/main/java/com/atm/dto/AuthenticatedUserDto.java
package com.atm.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
public class AuthenticatedUserDto {
    private String userName;
    private String password;


}